risk_ranger/requirements.txt
streamlit==1.24.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.21
python-dotenv==1.0.0
fpdf==1.7.2
colorama==0.4.6
pandas==2.1.1
openpyxl==3.1.2
uuid==1.30
logging==0.5.1.2